requirements.txt
streamlit
openai
psycopg2-binary
pgvector
pypdf
python-docx
beautifulsoup4
lxml
python-dotenv
fastapi
uvicorn[standard]
requests